Fix delay for reading serial numbe
It has been a while since working on this driver for the last time. If I
remember correctly, reading the address already worked at some point in
time. But now this is not the case either directly connected to a
nRF52840-DK or behind a TCA9548A mux.

Nonetheless, the time penalty should be neglectable in this case.